requirements.txt
fastapi
uvicorn[standard]
python-jenkins
pydantic-settings
